+++ hystrix/src/main/java/org/seedstack/netflix/hystrix/internal/HystrixCommandAnnotationResolver.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.netflix.hystrix.internal.guice;
+package org.seedstack.netflix.hystrix.internal;
 
 import org.seedstack.netflix.hystrix.internal.annotation.HystrixCommand;
 import org.seedstack.shed.reflect.StandardAnnotationResolver;