+++ hystrix/src/main/java/org/seedstack/netflix/hystrix/internal/HystrixPlugin.java
@@ -5,18 +5,20 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.netflix.hystrix.internal.guice;
 
+package org.seedstack.netflix.hystrix.internal;
+
+import com.netflix.hystrix.strategy.HystrixPlugins;
 import io.nuun.kernel.api.plugin.InitState;
 import io.nuun.kernel.api.plugin.context.InitContext;
 import io.nuun.kernel.api.plugin.request.ClasspathScanRequest;
-import org.kametic.specifications.Specification;
-import org.seedstack.netflix.hystrix.internal.annotation.HystrixCommand;
-import org.seedstack.seed.core.internal.AbstractSeedPlugin;
-
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Map;
+import org.kametic.specifications.Specification;
+import org.seedstack.netflix.hystrix.HystrixConfig;
+import org.seedstack.netflix.hystrix.internal.annotation.HystrixCommand;
+import org.seedstack.seed.core.internal.AbstractSeedPlugin;
 
 public class HystrixPlugin extends AbstractSeedPlugin {
     private Collection<Class<?>> scannedClasses = new ArrayList<>();
@@ -24,7 +26,7 @@ public class HystrixPlugin extends AbstractSeedPlugin {
 
     @Override
     public String name() {
-        return "hystrix";
+        return "netflix-hystrix";
     }
 
     @Override
@@ -36,9 +38,12 @@ public class HystrixPlugin extends AbstractSeedPlugin {
 
     @Override
     protected InitState initialize(InitContext initContext) {
-        Map<Specification, Collection<Class<?>>> scannedClassesBySpecification = initContext.scannedTypesBySpecification();
+        Map<Specification, Collection<Class<?>>> scannedClassesBySpecification = initContext
+                .scannedTypesBySpecification();
         scannedClasses.addAll(scannedClassesBySpecification.get(hystrixCommandSpecification));
-        return super.initialize(initContext);
+        ((CoffigHystrixDynamicProperties) HystrixPlugins.getInstance().getDynamicProperties())
+                .setHystrixConfig(getConfiguration(HystrixConfig.class));
+        return InitState.INITIALIZED;
     }
 
     @Override