+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 1.0.2 (2017-11-31)
+
+* [new] [Hystrix properties](https://github.com/Netflix/Hystrix/wiki/Configuration#execution.isolation.thread.timeoutInMilliseconds) can now be set directly in SeedStack configuration. 
+* [chg] Updated Hystrix to 1.5.13
+* [chg] Remove unneeded dependency to Netflix Archaius in Hystrix module.  
+
 # Version 1.0.1 (2017-05-02)
 
 * [chg] Feign has been moved to its own [add-on](https://github.com/seedstack/feign-addon).